+++ packages/wonder-blocks-button/components/button.js
@@ -2,7 +2,7 @@
 import * as React from "react";
 import PropTypes from "prop-types";
 
-import {getClickableBehavior} from "@khanacademy/wonder-blocks-core";
+import {getClickableBehavior} from "@khanacademy/wonder-blocks-clickable";
 import type {AriaProps, StyleType} from "@khanacademy/wonder-blocks-core";
 import type {IconAsset} from "@khanacademy/wonder-blocks-icon";
 import ButtonCore from "./button-core.js";
