+++ packages/wonder-blocks-clickable/index.js
@@ -1,4 +1,14 @@
 // @flow
+import type {
+    ClickableHandlers,
+    ClickableState,
+    ClickableRole,
+} from "./components/clickable-behavior.js";
 import Clickable from "./components/clickable.js";
 
+export {default as ClickableBehavior} from "./components/clickable-behavior.js";
+export {default as getClickableBehavior} from "./util/get-clickable-behavior.js";
+
 export {Clickable as default};
+
+export type {ClickableHandlers, ClickableState, ClickableRole};
