+++ packages/wonder-blocks-dropdown/components/select-opener.js
@@ -7,7 +7,8 @@ import PropTypes from "prop-types";
 import type {AriaProps} from "@khanacademy/wonder-blocks-core";
 
 import Color, {mix, fade} from "@khanacademy/wonder-blocks-color";
-import {addStyle, getClickableBehavior} from "@khanacademy/wonder-blocks-core";
+import {addStyle} from "@khanacademy/wonder-blocks-core";
+import {getClickableBehavior} from "@khanacademy/wonder-blocks-clickable";
 import Icon, {icons} from "@khanacademy/wonder-blocks-icon";
 import Spacing from "@khanacademy/wonder-blocks-spacing";
 import {LabelMedium} from "@khanacademy/wonder-blocks-typography";
