+++ packages/wonder-blocks-icon-button/components/icon-button.js
@@ -2,7 +2,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import {getClickableBehavior} from "@khanacademy/wonder-blocks-core";
+import {getClickableBehavior} from "@khanacademy/wonder-blocks-clickable";
 import type {IconAsset} from "@khanacademy/wonder-blocks-icon";
 import type {AriaProps, StyleType} from "@khanacademy/wonder-blocks-core";
 import IconButtonCore from "./icon-button-core.js";
