+++ styleguide.config.js
@@ -41,7 +41,6 @@ module.exports = {
             content: "packages/wonder-blocks-core/docs.md",
             components: [
                 "packages/wonder-blocks-core/util/add-style.js",
-                "packages/wonder-blocks-core/components/clickable-behavior.js",
                 "packages/wonder-blocks-core/components/id-provider.js",
                 "packages/wonder-blocks-core/util/server.js",
                 "packages/wonder-blocks-core/components/text.js",
